% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool1.R
\name{cellex_analysis}
\alias{cellex_analysis}
\title{Perform Enrichment Analysis}
\usage{
cellex_analysis(
  input_set,
  input_type = c("ensembl", "uniprot", "gene"),
  cellex_data = c(1, 2),
  is_cellex_compressed = TRUE,
  first_order = FALSE,
  all_genes_as_background = FALSE,
  statistic = c("W", "KS", "T"),
  p_value = TRUE,
  p_value_adjust = TRUE,
  emp_p_value = FALSE,
  es_value = FALSE,
  statistic_plot = FALSE,
  reps = 1000,
  num_cores = 1,
  num_background_genes = 0,
  save_output = TRUE
)
}
\description{
This function takes as input a gene set (in ENSEMBL, UNIPROT, or HNGC format) and performs
enrichment analysis / gene specificity analysis and outputs a dataframe with p-values and/or
scores indicating how specifically expressed the gene set is in different cell types.
}
