% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool1.R
\name{get_alias}
\alias{get_alias}
\title{gene set specificity}
\usage{
Takes CELLEX data and an input gene set (in ensembl, uniprot, or hgnc format),
then extracts the ensembl gene names and performs a user-specified test to determine
gene specific expression for various cell types.
}
\value{
Output dataframe with p-values, z-values, and/or enrichment scores (es-values)
for the various cell types.
}
\description{
gene set specificity
}
\author{
Mads Porse Pedersen
}
