% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellex_gsea.R
\name{get_alias}
\alias{get_alias}
\title{CELLEX-GSEA}
\usage{
Takes CELLEX data and an input gene set (in ensembl, uniprot, or hgnc format),
then extracts the ensembl gene names and performs a user-specified statistical test
to determine gene specific expression for various cell types.
}
\value{
Outputs dataframe with either empirical p-values and z-values, p-values
(non-empirical) or enrichment score (es) values for the various cell types.
}
\description{
CELLEX-GSEA
}
